requirements.txt
PyQt5>=5.15.0
Pillow>=8.0.0
python-dateutil>=2.8.0
watchdog>=2.1.0